+++ src/domain/api/client.ts
@@ -1,25 +1,55 @@
-import ApolloClient from 'apollo-boost';
-import { toast } from 'react-toastify';
+import { ApolloClient } from 'apollo-client';
+import { createHttpLink } from 'apollo-link-http';
+import { InMemoryCache } from 'apollo-cache-inmemory';
+import { setContext } from 'apollo-link-context';
+import { onError } from 'apollo-link-error';
+import { ApolloLink } from 'apollo-link';
+import * as Sentry from '@sentry/browser';
 
-import i18n from '../../common/translation/i18n/i18nInit';
-import { store } from '../app/state/AppStore';
 import { apiTokenSelector } from '../auth/state/AuthenticationSelectors';
-
-export default new ApolloClient({
-  request: async operation => {
-    try {
-      const token = apiTokenSelector(store.getState());
-      operation.setContext({
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      });
-    } catch (e) {
-      // User not authenticated
-      toast(i18n.t('api.errorMessage'), {
-        type: toast.TYPE.ERROR,
-      });
-    }
-  },
+import { store } from '../app/state/AppStore';
+import { showExpiredSessionPrompt } from '../app/state/ui/UIActions';
+import { fetchTokenError } from '../auth/state/BackendAuthenticationActions';
+const httpLink = createHttpLink({
   uri: process.env.REACT_APP_API_URI,
 });
+
+const errorLink = onError(({ graphQLErrors, networkError }) => {
+  if (graphQLErrors) {
+    graphQLErrors.forEach(({ message, locations, path }) =>
+      Sentry.captureMessage(
+        `[GraphQL error]: Message: ${message}, Location: ${locations}, Path: ${path}`
+      )
+    );
+  }
+
+  if (networkError) {
+    Sentry.captureMessage('Network error');
+  }
+  // Now show session timeout for all kind of error.
+  // In near future when backend comeup with better
+  // statusCode from error body
+  // TODO: Error can be handle better here.
+  store.dispatch(showExpiredSessionPrompt());
+
+  // Clear old token in favor of avoiding Apollo loop
+  store.dispatch(
+    fetchTokenError({ message: 'Token expired', name: 'fetchTokenError' })
+  );
+});
+const authLink = setContext((_, { headers }) => {
+  const token = apiTokenSelector(store.getState());
+  return {
+    headers: {
+      ...headers,
+      authorization: token ? `Bearer ${token}` : null,
+    },
+  };
+});
+
+const client = new ApolloClient({
+  link: ApolloLink.from([errorLink, authLink, httpLink]),
+  cache: new InMemoryCache(),
+});
+
+export default client;