+++ src/domain/api/generatedTypes/deleteChild.ts
@@ -1,0 +1,21 @@
+/* tslint:disable */
+/* eslint-disable */
+// This file was automatically generated and should not be edited.
+
+import { DeleteChildMutationInput } from "./globalTypes";
+
+// ====================================================
+// GraphQL mutation operation: deleteChild
+// ====================================================
+
+export interface deleteChild_deleteChild {
+  clientMutationId: string | null;
+}
+
+export interface deleteChild {
+  deleteChild: deleteChild_deleteChild | null;
+}
+
+export interface deleteChildVariables {
+  input: DeleteChildMutationInput;
+}