+++ src/domain/app/header/userDropdown/UserDropdown.tsx
@@ -1,19 +1,16 @@
 import * as React from 'react';
 import { useTranslation } from 'react-i18next';
-import { useSelector, useDispatch } from 'react-redux';
+import { useSelector } from 'react-redux';
 import { useHistory } from 'react-router-dom';
+import { useMatomo } from '@datapunt/matomo-tracker-react';
 
 import personIcon from '../../../../assets/icons/svg/person.svg';
 import Dropdown from '../../../../common/components/dropdown/Dropdown';
 import { isAuthenticatedSelector } from '../../../auth/state/AuthenticationSelectors';
 import { profileSelector } from '../../../profile/state/ProfileSelectors';
 import { loginTunnistamo, logoutTunnistamo } from '../../../auth/authenticate';
-import { resetFormValues } from '../../../registration/state/RegistrationActions';
-import { persistor } from '../../state/AppStore';
-import { clearProfile } from '../../../profile/state/ProfileActions';
-import { resetBackendAuthentication } from '../../../auth/state/BackendAuthenticationActions';
 import UserMenu from '../userMenu/UserMenu';
-import client from '../../../api/client';
+import { flushAllState } from '../../../auth/state/AuthenticationUtils';
 
 export interface UserDropdownProps {
   isSmallScreen?: boolean;
@@ -26,23 +23,18 @@ const UserDropdown: React.FunctionComponent<UserDropdownProps> = ({
   const history = useHistory();
   const isAuthenticated = useSelector(isAuthenticatedSelector);
   const profileData = useSelector(profileSelector);
-  const dispatch = useDispatch();
+  const { trackEvent } = useMatomo();
 
   const logout = {
     label: t('authentication.logout.text'),
     id: 'logoutButton',
     onClick: () => {
-      // Clear user form data
-      dispatch(resetFormValues());
-      // Clear profile (fetched from API)
-      dispatch(clearProfile());
-      // Clear backend auth data
-      dispatch(resetBackendAuthentication());
-      // Flush data in redux store and localStorage
-      persistor.flush();
-      // Clear Apollo cache
-      client.clearStore();
-      // Log out
+      trackEvent({ category: 'action', action: 'Log out' });
+
+      // Flush all cached state
+      flushAllState({});
+
+      // Log out;
       logoutTunnistamo();
     },
   };
@@ -71,7 +63,10 @@ const UserDropdown: React.FunctionComponent<UserDropdownProps> = ({
     id: 'loginButton',
     label: t('authentication.login.text'),
     icon: personIcon,
-    onClick: () => loginTunnistamo(),
+    onClick: () => {
+      trackEvent({ category: 'action', action: 'Log in' });
+      loginTunnistamo();
+    },
   };
 
   if (!isSmallScreen) {