+++ src/domain/auth/state/AuthenticationSelectors.ts
@@ -1,5 +1,8 @@
 import { StoreState } from '../../app/types/AppTypes';
 
+export const userAccessTokenSelector = (state: StoreState) =>
+  state.authentication.tunnistamo.user?.access_token;
+
 export const apiTokenSelector = (state: StoreState) =>
   state.authentication.backend.apiToken;
 
@@ -8,7 +11,7 @@ export const isLoadingUserSelector = (state: StoreState) =>
   state.authentication.backend.isFetchingToken;
 
 export const isAuthenticatedSelector = (state: StoreState) =>
-  state.authentication.backend.isAuthenticated;
+  !!state.authentication.backend.apiToken;
 
 export const userSelector = (state: StoreState) =>
   state.authentication.tunnistamo.user;